[Refactor] UIColor, CGRect 변환을 VC에서 하도록 변경 - 하위 Model에서는 UIKit을 import하지 않도록함 - UIColor와 CGRect를 구성할 수 있는 값만 튜플로 묶어서 전달, 변환은 VC가 담당
(cherry picked from commit df03c58957c707889d39313902e48090df1e895c)
(cherry picked from commit b555b9f08e5c0151bc5d4538bc1bc2adcd34d34c)

+++ DrawingApp/DrawingApp/Views/SideInspectorView.swift
@@ -166,7 +166,7 @@ class SideInspectorView: UIStackView {
         backgroundColorValueButton.backgroundColor = .systemGray5
     }
     
-    func setAlphaValueLabelText(by text: Float) {
+    func setAlphaValueLabelText(by text: Double) {
         alphaValueLabel.text = "\(text)"
     }
